update library
update pagination, movie-card markup, waiting for modal buttons
