+++ src/library.html
@@ -16,6 +16,6 @@
   <body>
     <include src="./partials/library-header.html"></include>
     <include src="./partials/library-main.html"></include>
-    <!-- <include src="./partials/footer.html"></include> -->
+    <include src="./partials/footer.html"></include>
   </body>
 </html>
