+++ src/partials/library-main.html
@@ -1,7 +1,7 @@
 <main class="main">
   <section class="pagination">
     <div class="container">
-      <ul class="pagination__movies" aria-live="polite"></ul>
+      <ul class="gallery" aria-live="polite"></ul>
       <nav class="pagination__wrapper">
         <button
           class="pagination__button"
@@ -9,8 +9,8 @@
           title="Previous page"
           aria-label="Previous page"
         >
-          <svg class="footer__heart">
-            <use href="./images/svg/sprite.svg#footer-heart"></use>
+          <svg class="pagination__arrow">
+            <use href="./images/svg/sprite.svg#arrow-left"></use>
           </svg>
         </button>
 
@@ -22,8 +22,8 @@
           title="Next page"
           aria-label="Next page"
         >
-          <svg class="footer__heart">
-            <use href="./images/svg/sprite.svg#footer-heart"></use>
+          <svg class="pagination__arrow">
+            <use href="./images/svg/sprite.svg#arrow-right"></use>
           </svg>
         </button>
       </nav>
