+++ src/sass/layouts/_library-pagination.scss
@@ -4,6 +4,8 @@
         display: flex;
         align-items: center;
         justify-content: center;
+
+        // padding-bottom: 140px;
     }
 
     &__number, &__button {
@@ -20,6 +22,11 @@
     &__number.active{
         color: #fff;
         background: $accent-text-color;
+    }
+
+    &__arrow {
+        height: 16px;
+        width: 16px;
     }
 
     &__pages span {
