+++ src/sass/utils/_base.scss
@@ -61,6 +61,7 @@ body {
       margin-right: auto;
       padding-right: 20px;
       padding-left: 20px;
+      margin-bottom: 80px;
   
       @media screen and (min-width: 320px) {
         width: 320px;
